requirement.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.2
djangorestframework==3.15.2
et_xmlfile==2.0.0
idna==3.10
kavenegar==1.1.2
openpyxl==3.1.5
pillow==11.0.0
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
